robot_arm: split up largest robot_arm.hpp file.
